requirements.txt
mcp>=0.9.0
uiautomator2>=3.0.0
httpx>=0.25.0
Pillow>=10.0.0
python-dotenv>=1.0.0
